Broker API handler on external http server
Currently, it wasn't possible to register the Broker APISurface
handler on an external HTTP server, since the registration was
included in the server.New() method that instantiates also the
server.Server.

This commit is splitting the current server.New(...) method in two
in order to have a specific method for APISurface handler
registration: server.registerHTTPHandler(...).

An additional Public method NewHTTPHandler(...) is added in order
to allow a user to register on an external HTTP server the Broker
API endpoints.